waste-wise/backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# GeoDjango and PostGIS
django-leaflet==0.29.0
djangorestframework-gis==1.0

# Authentication
djangorestframework-simplejwt==5.3.0
djoser==2.2.2

# WebSocket and Real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# MQTT
paho-mqtt==1.6.1

# Celery
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Redis
redis==5.0.1
django-redis==5.4.0

# Utilities
python-decouple==3.8
Pillow==10.1.0
django-filter==23.4
django-extensions==3.2.3

# API Documentation
drf-yasg==1.21.7

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
ipython==8.17.2
django-debug-toolbar==4.2.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Monitoring and Logging
sentry-sdk==1.38.0
django-prometheus==2.3.1

# Security
django-ratelimit==4.1.0
django-guardian==2.4.0

# File handling
whitenoise==6.6.0

# Timezone support
pytz==2023.3.post1